[Add & Update] Add directoryMode, new Test, update README
Add directoryMode, which requires the directories `original` and `reproduced` as (direct) subdirectories of the directory path passed as parameter in config JSON.
Add Test, including test directory, and a clean-up function for the test.
Minor bugfixes, especially in error handling. Reject messages could at times not be handled correctly, as an Array was expected in the reject handler function in `index.js`.
Minor improvements.

Updated README. Recent updates are now described.
Minor additions in commentary.